refactor: перенёс ProtectedRoute в pages
